Implement file separation for recipes #68
