+++ practice/src/index.html
@@ -8,7 +8,7 @@
   <link rel="stylesheet" href="./styles/index.scss" />
   <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.0.0/dist/css/bootstrap.min.css"
     integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous">
-  <title>Food Blog</title>
+  <title>Recipes</title>
 </head>
 
 <body>
@@ -181,21 +181,23 @@
       <!-- Item-food 4 -->
       <ul class="delicious__list">
         <li class="delicious__item">
-          <article class="card">
-            <img class="card__image delicious__image"
-              src="./assets/images/introduce-food-item/caremel-strawberry-milkshake.png"
-              alt="Picture caremel-strawberry-milkshake" />
-            <div class="card__rating">
-              <span class="card__rating-star"></span>
-              <span class="card__rating-star"></span>
-              <span class="card__rating-star"></span>
-              <span class="card__rating-star"></span>
-              <span class="card__rating-star"></span>
-            </div>
-            <p class="card__name delicious__name-food">
-              Caramel Strawberry Milkshake
-            </p>
-          </article>
+          <a href="recipe.html">
+            <article class="card">
+              <img class="card__image delicious__image"
+                src="./assets/images/introduce-food-item/caremel-strawberry-milkshake.png"
+                alt="Picture caremel-strawberry-milkshake" />
+              <div class="card__rating">
+                <span class="card__rating-star"></span>
+                <span class="card__rating-star"></span>
+                <span class="card__rating-star"></span>
+                <span class="card__rating-star"></span>
+                <span class="card__rating-star"></span>
+              </div>
+              <p class="card__name delicious__name-food">
+                Caramel Strawberry Milkshake
+              </p>
+            </article>
+          </a>
         </li>
 
         <!-- Item food 5 -->
@@ -241,7 +243,7 @@
 
     <!-- START CATEGORIES  -->
     <section class="container category">
-      <h3 class="category__title">Popular Categories</h3>
+      <h3 class="category__title delicious__title">Popular Categories</h3>
       <ul class="category__list">
         <li class="category__item">
           <article class="card-category">
@@ -378,16 +380,18 @@
 
     <!-- START RECIPES -->
     <section class="container recipes">
-      <h4 class="recipes__title">Latest Recipes</h4>
+      <h4 class="recipes__title delicious__title">Latest Recipes</h4>
       <ul class="recipes__list">
-        <li class="recipes__food">
-          <article class="card">
-            <img class="card__image" src="./assets/images/recipes/caramel-strawberry-milkshake.png"
-              alt="Pictire caramel-strawberry-milkshake" />
-            <p class="card__name recipes__name">
-              Caramel Strawberry Milkshake
-            </p>
-          </article>
+        <!-- <li class="recipes__food">
+          <a href="recipe.html">
+            <article class="card">
+              <img class="card__image" src="./assets/images/recipes/caramel-strawberry-milkshake.png"
+                alt="Pictire caramel-strawberry-milkshake" />
+              <p class="card__name recipes__name">
+                Caramel Strawberry Milkshake
+              </p>
+            </article>
+          </a>
         </li>
         <li class="recipes__food">
           <article class="card">
@@ -571,7 +575,7 @@
               Four Ingredient Oatmeal Pancakes
             </p>
           </article>
-        </li>
+        </li> -->
       </ul>
       <div class="recipes__load">Load More</div>
     </section>
@@ -646,7 +650,7 @@
     </div>
   </footer>
   <!-- END FOOTER -->
-
+  <script type="module" src="./javascript/app.js"></script>
 </body>
 
 </html>
