+++ practice/src/javascript/controllers/toggle-controller.js
@@ -1,0 +1,10 @@
+export default class ToggleController {
+  constructor( view) {
+    this.view = view;
+  }
+
+  init = async () => {
+    this.view.bindCallback("menuToggle");
+  };
+}
+
