+++ practice/src/javascript/recipes-app.js
@@ -1,0 +1,6 @@
+import RecipesAppView from "./view/recipes-app/recipes-view";
+
+window.addEventListener("load", () => {
+const recipeAppController = new recipeAppController(new RecipesAppView())
+recipeAppController.init();
+});
