+++ practice/src/javascript/view/recipes-app/recipes-view.js
@@ -1,0 +1,21 @@
+export default class RecipesAppView {
+  constructor() {
+    // Render recipes
+    this.recipesListEl = document.querySelector(".recipes__list");
+  }
+
+  /**
+   * The bindCallback function binds events to corresponding handlers on HTML elements in the user interface.
+   * @param {string} event - Event name to attach.
+   * @param {function} handler - The handler function is called when the event occurs.
+   */
+  bindCallback = (event, handler) => {
+    switch (event) {
+      
+      default:
+        break;
+    }
+  };
+
+
+}
