+++ practice/src/user-manager.html
@@ -6,7 +6,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   <link rel="icon" href="./assets/images/logo/icon-logo.png" type="image/x-icon" />
   <link rel="stylesheet" href="./styles/index.scss" />
-  <title>User manager-Food Blog</title>
+  <title>Dashboard</title>
 </head>
 
 <body>
@@ -15,7 +15,7 @@
     <div class="navbar-user__wrapper">
       <div class="navbar-user__item">
         <span class="navbar-user__icon-menu"></span>
-        <h1 class="navbar-user__title">User Manager</h1>
+        <h1 class="navbar-user__title">Dashboard</h1>
       </div>
     </div>
   </header>
